fix: ensure proper casting Redis env variables on PHP 8.5

+++ config/database.php
@@ -157,8 +157,8 @@ return [
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'username' => env('REDIS_USERNAME'),
             'password' => env('REDIS_PASSWORD'),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_DB', 0),
+            'port' => (int) env('REDIS_PORT', 6379),
+            'database' => (int) env('REDIS_DB', 0),
         ],
 
         'cache' => [
@@ -166,8 +166,8 @@ return [
             'host' => env('REDIS_HOST', '127.0.0.1'),
             'username' => env('REDIS_USERNAME'),
             'password' => env('REDIS_PASSWORD'),
-            'port' => env('REDIS_PORT', 6379),
-            'database' => env('REDIS_CACHE_DB', 1),
+            'port' => (int) env('REDIS_PORT', 6379),
+            'database' => (int) env('REDIS_CACHE_DB', 1),
         ],
 
     ],
